+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixed
 - crash when a member gets added to a group and "View Group" dialog is open #5111
 - show appropriate state in AddMemberDialog #5114
+- tauri: accessibility: fix focus always being locked on the message input
 
 <a id="1_58_2"></a>